Fix README: Twitter no longer uses Jobvite
Thanks for this repository!

I tried the example from the README and it was failing due to the
fact that Twitter no longer employs Jobvite for job applications.

I replaced the company ID of Twitter with another company that is
still active on Jobvite (TiVo) and the code worked flawlessly.